refactor(samplers): stream leaf calls with caller-owned collection
Drive sampling and replay through one generator executor, keeping physical microbatching in adapters. Share cold-start dummy inputs collectively and carry their gradients across variants. Restore per-evaluation state isolation, group pairwise completions by prompt key, and keep sampler and loss arithmetic in fp32.

Restore caller collectors and serving step progress, migrate training callers and configs, and update execution documentation. Preserve 21 numeric fixtures; validate 70 tests, 154 subtests, two-rank Gloo workers, and two-GPU FSDP and bf16 tiled sampling.

+++ flow_control/processors/base.py
@@ -238,7 +238,7 @@ class BaseProcessor[
             (1, c, h, w), generator=generator, device=device, dtype=dtype
         )
         if self.initial_noise_scale != 1.0:
-            latents = latents * self.initial_noise_scale
+            latents = (latents.float() * self.initial_noise_scale).to(dtype)
         noisy_latents = batch["noisy_latents"] = self._pack_latents(latents)
         return noisy_latents
